# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EcoLoop",
        "description": "Application de dons et d'échanges d'objets durables",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Categorie"
                ],
                "summary": "liste de toutes les categories",
                "operationId": "49c0c4231d78a88f15f6c79647f3b578",
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categorie"
                ],
                "summary": "Ajout d'une categorie",
                "operationId": "af9c0d174b0b4abc44d74d3815c2f2af",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "example": "libelle"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categorie créée avec succées"
                    },
                    "401": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/nombreCategories": {
            "get": {
                "tags": [
                    "Categorie"
                ],
                "summary": "Nombres de categories",
                "operationId": "fe079482f1931182b62a3e2eaa4cf712",
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categories/{categorie}": {
            "put": {
                "tags": [
                    "Categorie"
                ],
                "summary": "Modifier la categorie",
                "operationId": "47563c52ddabd478aa02b2073a6601cc",
                "parameters": [
                    {
                        "name": "Categorie",
                        "in": "path",
                        "description": "ID de la categorie à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    " libelle": {
                                        "type": "string",
                                        "example": "Nouveau libelle"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categorie"
                ],
                "summary": "Supprimer une categorie",
                "operationId": "5ec2cdce22ea1866e6fc07793ef7553b",
                "parameters": [
                    {
                        "name": "Categorie",
                        "in": "path",
                        "description": "ID de la categorie à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/commentaires/{article}": {
            "get": {
                "tags": [
                    "Commentaire"
                ],
                "summary": "liste de tous les commentaires",
                "operationId": "b3537d09e5619d47f8c90f0eac007a4a",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "ID de l'article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Commentaire"
                ],
                "summary": "Ajout d'un commentaire",
                "operationId": "4ddb3858cda0a8a02d939c0e9e915974",
                "parameters": [
                    {
                        "name": "article",
                        "in": "path",
                        "description": "ID de l'article à commenter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "commentaire": {
                                        "type": "string",
                                        "example": "commentaire"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commentaire ajouté avec succées"
                    },
                    "401": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/commentaires/{commentaire}": {
            "delete": {
                "tags": [
                    "Commentaire"
                ],
                "summary": "Supprimer un commentaire",
                "operationId": "02c1e1681acd5dfe6ecf304fc1e60ecf",
                "parameters": [
                    {
                        "name": "commentaire",
                        "in": "path",
                        "description": "ID du commentaire à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/votes/{user}": {
            "get": {
                "tags": [
                    "Evaluation"
                ],
                "summary": "Le nombre de vote d'un utilisateur donné",
                "operationId": "1bb421732c9c8ded1829cdd8f9ac7270",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succes"
                    }
                }
            }
        },
        "/api/voter/{user}": {
            "post": {
                "tags": [
                    "Evaluation"
                ],
                "summary": "Voter pour un utilisateur",
                "operationId": "797a44ce26f3de32616edcb27466bd5c",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Vote effectué avec succées"
                    },
                    "401": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}